requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
sqlalchemy
psycopg2-binary
pgvector
faiss-cpu
sentence-transformers
python-dotenv
scikit-learn
asyncpg
numpy
python-multipart
twilio
pytest 
pytest-asyncio 
httpx